Gemini/PARIS/requirements.txt
flask
playwright
gunicorn
selenium
python-dotenv